fix: update CameraSelectableItems.cs filename
